Differentiate %pyproject_save_files based on the automode setting
For automated environments, we want to ensure buildability of the
created specfile, hence adding `"*" +auto` to %pyproject_save_files.
In the real Fedora packages, the valid argument to %pyproject_save_files
is a list of module names.
For now there's no automatic way to discover what modules the package
contains, so when a human user invokes pyp2spec, we want to generate a
file with placeholder (...) for them to fill in after the package
inspection.