Move RPM-specific computations to conf2spec
pyp2conf shouldn't be so heavy with RPM-specific data.
The data stored in the config file are the ones read from either the
PyPI API or Metadata file.
Everything related to transformations to specfiles is now stored in
conf2spec. It reads the config data and converts them to RPM-specific
format.
Ideally, if pyproject-rpm-macros become obsolete and a new system comes
in place, the only changes would be applicable on the side of conf2spec.